+++ docs/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.4.0
 
-* See [140UPGRADE](140UPGRADE.md) for upgrade instructions
+* See [140upgrade.md](140upgrade.md) for upgrade instructions
 * Switch to sqlalchemy for data access
 * Refactor to allow pluggable storage backend